Correcting image resizing using the UIImage+Resize
The current Rhodes image resizing does not work correctly. It results in rotated images.

https://groups.google.com/forum/?fromgroups#!topic/rhomobile/W0b46OllwRw

The edits done here using the UIImage+Resize class extension to fix the problem. Add the following files to Rhodes:

https://github.com/mbcharbonneau/UIImage-Categories

There is one small bug in them. You need to comment out the check for iOS5 in the resizeImage function:

Comment out:

    if ( [[[UIDevice currentDevice] systemVersion] floatValue] >= 5.0 ) 
    {
        drawTransposed = NO;  
    } 
    else 
    {   