Make rho_sync_issyncing call issyncing() and also check if m_pImpl is NULL.
In RhoConnect, issyncing calls rho_sync_issyncing (a C function it seems), which works on the version of LLVM I have installed with Xcode 4.x, but the LLVM packaged with Xcode 5 tries to call rho_sync_issyncing as a class method. That class method then calls rho_sync_issyncing, which it interprets as a class method rather than the C function, so it gets into a loop where it eventually crashes.

This fix WILL NOT completely fix this bug. You will also have to modify Classes/impl/initRhoConnectClient.cpp in the RhoConnect client gem and comment out the class definition of rho_sync_issyncing (probably also need to do it in the header).